Implements proper lengths for columns.
Introduced the Field struct which holds, in a linear list, informations about
each field. The item fields has been renamed to pieces in order to prevent
confusion. Computation of column lengths is done into the getmaxlength()
function and are passed as array to the items and fields drawing routines.

All of the above required a bit of refactoring. This commit also fixes some bug
and reimplements cleanupitems() which was totally crap.