Remove cloneitem().
By thinking further, we may have a widget with a list to switch between views,
in this case it's needed to clone the selected item in order to mantain the
data, regardless of the state of the previous view. But for now we're stuck
with a views hierachy, so it's safe to only take a reference to the selected
item of the previous view, because it cannot be destroyed prematurely.