Edit records feature.
Records in tables that are missing an unique key are not editables. Some minor
improvement has also been added, like the improvement of mysql_exec() which
don't care about sql size anymore due to the introduction of vasprintf().

The code has to be cleaned up and better names for some function must be found.