Improve how the $EDITOR gets called.
First we call endwin() (which is called by stfl_free() through ui_end()),
then take off all the signal handlers from ncurses. At this point we can
fork the editor and wait or it to exit. By calling ui_init() we restore
ncurses signal handlers and may safely continue execution.

This is quite simple but required me to dig in some code. For curious, give a
look at the mutt's _mutt_system() function.

As a side effect, it was also possible to clean up fget() which was weirdly
removing a newline from the buffer.