Refactoring records editing.
The big change is the ui_sql_edit_exec() routine which is reusable and allow to
edit a file, then exec the content into MySQL (if modified). This comes with a
bit of refactoring. In case of wrong SQL code the current file may be
re-opened. Also flow control chars are now escaped so that they are not
interpreted by MySQL.

This commit is a significant step further in order to write routines like, for
example, edittable() in a very simple manner. Though, it raises a strong need
to cleanup the code.